+++ SF3Lib/Models/Structs/MPD/SurfaceModel/TileTextureRow.cs
@@ -22,7 +22,7 @@ namespace SF3.Models.Structs.MPD.SurfaceModel {
             // TODO: somehow cache this?
             var tiles = new ushort[64];
             for (var i = 0; i < tiles.Length; i++)
-                tiles[i] = (ushort) Data.GetUInt16(xAddress[i]);
+                tiles[i] = Data.GetUInt16(xAddress[i]);
             return tiles;
         }
 
@@ -54,7 +54,7 @@ namespace SF3.Models.Structs.MPD.SurfaceModel {
             => SetTextureFlags(x, (byte) (GetTextureFlags(x) & ~0x80 | (value ? 0x80 : 0x00)));
 
         public ushort this[int index] {
-            get => (ushort) Data.GetUInt16(xAddress[index]);
+            get => Data.GetUInt16(xAddress[index]);
             set => Data.SetUInt16(xAddress[index], value);
         }
