+++ SF3Lib/Models/Structs/MPD/TextureIDStruct.cs
@@ -13,7 +13,7 @@ namespace SF3.Models.Structs.MPD {
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_textureIdAddr), displayName: "Texture ID", displayOrder: 0, displayFormat: "X2")]
         public ushort TextureID {
-            get => (ushort) Data.GetUInt16(_textureIdAddr);
+            get => Data.GetUInt16(_textureIdAddr);
             set => Data.SetUInt16(_textureIdAddr, value);
         }
     }
